+++ piet/src/null_renderer.rs
@@ -161,7 +161,7 @@ impl TextLayout for NullTextLayout {
         HitTestPoint::default()
     }
 
-    fn hit_test_text_position(&self, _text_position: u32, _trailing: bool) -> Option<HitTestTextPosition> {
+    fn hit_test_text_position(&self, _text_position: usize, _trailing: bool) -> Option<HitTestTextPosition> {
         None
     }
 }
